+++ src/app/questions/categories/category-detail/category-detail.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Router } from '@angular/router';
 import { DataService } from 'src/app/data.service';
 import { Category } from 'src/app/model/Category';
@@ -12,6 +12,9 @@ export class CategoryDetailComponent implements OnInit {
 
   @Input()
   category: Category;
+  @Output()
+  dataChangedEvent = new EventEmitter();
+  message = '';
 
   constructor(private router: Router,
             private dataService: DataService) { }
@@ -25,10 +28,17 @@ export class CategoryDetailComponent implements OnInit {
   }
 
   deleteCategory(): void {
-    this.dataService.deleteCategory(this.category.id).subscribe(
-      next => {
-        this.router.navigate(['questions', 'categories']);
-      }
-    )
+    const deleteConfirmed = confirm('Are you sure you wish do delete this Category?');
+    if (deleteConfirmed) {
+      this.message = "Deleting...";
+      this.dataService.deleteCategory(this.category.id).subscribe(
+        next => {
+          this.dataChangedEvent.emit();
+          this.router.navigate(['questions', 'categories']);
+        }, error => {
+          this.message = 'Sorry, this Category cannot be deleted at this time';
+        }
+      )
+    }
   }
 }
